requirements.txt
setuptools
dateparser
pytz
python-dateutil
beautifulsoup4
BeautifulSoup
datefinder
timeout-decorator